requirements.txt
# Expert Panel Simulator Dependencies

# Core framework
pyautogen==0.2.35

# API providers (choose one or both)
openai>=1.3.0
anthropic>=0.18.0

# Environment and configuration
python-dotenv>=1.0.0
pyyaml>=6.0

# Document processing
markdown>=3.5
beautifulsoup4>=4.12.0

# Utilities
click>=8.1.0
rich>=13.5.0
tqdm>=4.66.0

# Optional: For advanced features
# tiktoken>=0.5.0  # Token counting for OpenAI
# numpy>=1.24.0   # For analytics